+++ salesforce/message.py
@@ -1,31 +1,11 @@
 SEPRATE = """
----------------------------------------------------------------------------------------------------
+-------------------------------------------------------------------------------
 {0}
----------------------------------------------------------------------------------------------------
+-------------------------------------------------------------------------------
 """
 
 INVALID_COMPONENT = "It's an invalid salesforce component"
 DELETE_CONFIRM_MESSAGE = "Are you sure you really want to delete this?"
 REFRESH_CONFIRM_MESSAGE = "Are you sure you really want to refresh this?"
 INVALID_NEW_COMPONENT_FORMAT = "invalid new component format"
-
-AUTHORIZATION_FAILED_MESSAGE = """
--------------------------------------------------------------------------
-Connecting to Salesforce Failed
--------------------------------------------------------------------------
-"""
-
-GENERATE_ALL_WORKBOOKS_REQUEST_LIMIT_MESSAGE = """
-Are you sure you really want to do it, this will consume your hundreds of API requests, more detail refer to
-http://login.salesforce.com/help/doc/en/integrate_api_rate_limiting.htm
-"""
-DOWNLOAD_ALL_SUCCESSFULLY = "Refresh All Successfully"
-
-# Metadata api part
-METADATA_CHECK = """
-You should click SublimeApex > Migration > Retrieve Metadata firstly
-"""
-
-initiate_sobject_completion = """
-Start initiate sobject completions.....
-"""
+METADATA_CHECK = "You should click SublimeApex > Migration > Retrieve Metadata firstly"